Refactor beam search
